Update Molecule tests to run on all supported platforms (#9)
The Molecule tests create a ZooKeeper and Kafka cluster consisting of three nodes running
within Docker containers. Each Docker container is a different OS to test the platforms
that this role supports.

Include Kafka role for testing and group and host vars. Group vars contain the global configuration
for ZooKeeper connection strings and the location of the Kafka bootstrap servers.

The inclusion of the Kafka role for testing uses an environment lookup to locate the project
directory of the role to be tested. This is to support being able to locate the role no matter
what the name of the directory it has been checked out to.

Update role metadata to list supported platforms.

Add group and user and file mode when creating directories and templating files.

Github Actions workflows added for Molecule tests and linting codebase:

- the Github super-linter is used for linting files
- the ansible-lint-action is used to lint the Ansible role files
- add status badges to README