Use LOG_DIR environment variable for application logs
The application logs are written to the /var/log/kafka directory and symlinked to the /opt/kafka/logs directory.

The value for the LOG_DIR environment variable for the process specifies where the application logs are written.
By default the /opt/kafka/logs directory is used if the LOG_DIR environment variable is not present.